requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
api==0.0.7
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.2
markdownify==0.11.6
MarkupSafe==2.1.2
multidict==6.0.4
nose==1.3.7
openai==0.27.4
outcome==1.2.0
pyChatGPT==0.4.3.3
pycparser==2.21
PySocks==1.7.1
requests==2.28.2
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
undetected-chromedriver==3.4.6
urllib3==1.26.15
websockets==11.0.1
Werkzeug==2.2.3
wincertstore==0.2
wsproto==1.2.0
yarl==1.8.2
